A few cool improvements!
- New option for inst: -strip.
- New option for inst: -dummy.
- Don't ask to remove .insfiles file if no changes needed.
What's more, the file removes "itself" on error.
- If a name for a file with list is chosen automatically,
don't start its name wih a dot (would be a hidden file on UNIX).
Of course this mechanism of creating a name (etc.) is still
a bit brain damaged :) and allows for ugly things (it's mentoned
in the README file), but that's not a big problem.
It may improve at some point; funcs like realpath will be the rescue.
- Another small yet nice changes to code and docs.